upgrade api v2

+++ app/src/main/java/com/bhw/covid_suburb_au/dashboard/viewmodel/BasicUiState.kt
@@ -26,6 +26,6 @@ data class StateItemViewObject(
     val stateCode: String,
     val stateFullName: String,
     val isMyState: Boolean = false,
-    val cases: Int,
+    val cases: Long,
     val isHighlighted: Boolean
 )
