+++ app/src/main/java/com/bhw/covid_suburb_au/dashboard/viewmodel/SuburbUiState.kt
@@ -3,9 +3,9 @@ package com.bhw.covid_suburb_au.dashboard.viewmodel
 
 data class SuburbUiState(
     val rank: Int,
-    val postcode: Long,
+    val postcode: Int,
     var briefName: String,
-    val cases: Int = 0,
+    val cases: Long = 0,
     val isHighlighted: Boolean = false,
     val isFollowed: Boolean = false,
     val isMySuburb: Boolean = false,
