+++ app/src/main/java/com/bhw/covid_suburb_au/data/helper/CovidDisplayHelper.kt
@@ -1,7 +1,7 @@
 package com.bhw.covid_suburb_au.data.helper
 
 object CovidDisplayHelper {
-    fun casesToDisplay(cases: Int) = if (cases > 0) {
+    fun casesToDisplay(cases: Long) = if (cases > 0) {
         "+$cases"
     } else cases.toString()
 }
