+++ app/src/main/java/com/bhw/covid_suburb_au/data/room/AuPostcodeDao.kt
@@ -9,10 +9,10 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface AuPostcodeDao {
     @Query("SELECT * FROM au_postcode WHERE postcode=:postcode")
-    suspend fun findPostcode(postcode: Long): AuPostcodeEntity?
+    suspend fun findPostcode(postcode: Int): AuPostcodeEntity?
 
     @Query("SELECT * FROM au_postcode WHERE postcode>=:low AND postcode<=:high")
-    suspend fun findPostcodesInRange(low: Long, high: Long): List<AuPostcodeEntity>
+    suspend fun findPostcodesInRange(low: Int, high: Int): List<AuPostcodeEntity>
 
     @Query("DELETE FROM au_postcode")
     suspend fun clear()
