+++ app/src/main/java/com/bhw/covid_suburb_au/settings/viewobject/SuburbMultipleSelectionListItem.kt
@@ -1,7 +1,7 @@
 package com.bhw.covid_suburb_au.settings.viewobject
 
 data class SuburbMultipleSelectionListItem(
-    val postcode: Long,
+    val postcode: Int,
     val display: String,
     val isSelectable: Boolean = true,
     var isSelected: Boolean = false
